requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
boto3==1.15.18
botocore==1.18.18
dj-database-url==0.5.0
Django==3.1.2
django-ckeditor==6.0.0
django-environ==0.4.5
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
phonenumbers==8.12.11
Pillow==8.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
